Converted the old composition into class
