+++ lib/index.js
@@ -6,10 +6,11 @@ var Typpy = require("typpy")
   , Composition = require("./composition")
   , Ul = require("ul")
   , EngineParser = require("engine-parser")
+  , Graph = require("./graph")
   ;
 
 class EngineBuilder extends EngineParser {
-    getGraph (level) {
+    getGraph (level, fn) {
 
         // Levels
         //  - 1 Big bubbles, connected based on events
@@ -22,7 +23,6 @@ class EngineBuilder extends EngineParser {
         };
 
         this.parse((err, parsed) => {
-
             debugger
         });
     }
